requirements.txt
absl-py==0.13.0
Flask==2.0.1
h5py==3.1.0
Jinja2==3.0.1
Keras-Preprocessing==1.1.2
numpy==1.19.5
pillow==8.2.0
python-dateutil==2.8.2
python-engineio==3.13.2
python-jwt==2.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0


